Update Serenity and Songbird
These were the necessary changes to get Songbird and serenity to play
nice with their new versions.

The major refactor here was the requirement to add our own queue for
Youtube tracks. This was because Songbird removed the attached Metadata
to track handles and I don't believe it's possible to retrieve this
information anymore. Hence the queue has been copied from the Songbird
source and modified to add a metadata member to the queued tracks.

Additionally, songbird now includes a yt-dlp integration, I've removed
the custom youtube Input type